Adds prototype of UWP gesture support
This is a *very* minimal viable product. Almost none of the gestures are constrained by the configuration (e.g., min/maxDelta for panning).

The approach uses the UWP GestureRecognizer API. The limitations of this approach are:
* There is no support for multi-tap beyond double tap
* The long press is not configurable
* The pan responder cannot set a minimum / maximum number of touch points
Some of these we can work around by using the pointer API directly instead of the GestureRecognizer API (e.g., taps).

This changeset doesn't include the NativeViewGestureHandler. Initial investigations makes it look like it could be quite difficult, as much of the emphasis in UWP is on "direct manipulation", which does not occur on the main UI thread and does not use the available pointer and manipulation APIs.

This changeset also does not include any attempt to implement gesture interactions (e.g., simultaneousHandlers and waitFor props).

Overall, this changeset barely scratches the surface for gesture handling, but is a useful starting point as some of the investigations above move forward.